+++ packages/client/modules/pages/Page.tsx
@@ -55,11 +55,11 @@ export const Page = (props: Props) => {
   // The editor is conditionally loaded only after syncing so the forced schema is not injected before
   // The yjs document loads
   return (
-    <div className='relative flex w-full flex-col items-center bg-white'>
+    <div className='relative flex flex-col items-center bg-white'>
       {isPublic ? <PageHeaderPublic /> : <PageHeader pageRef={page} />}
       <div
         className={cn(
-          'relative flex min-h-screen w-full justify-center bg-white pt-28 pb-10',
+          'relative flex min-h-screen w-full justify-center bg-white pt-28 pb-10 print:pt-0 print:caret-transparent',
           isDatabase ? 'max-w-9/10' : 'max-w-[960px]'
         )}
       >