+++ packages/server/database/types/User.ts
@@ -18,6 +18,7 @@ interface Input {
   pseudoId?: string | null
   sendSummaryEmail?: boolean
   tms?: string[]
+  persistentUserId?: string
 }
 
 export default class User {
@@ -40,6 +41,7 @@ export default class User {
   tms: string[]
   reasonRemoved?: string
   rol?: AuthTokenRole.SUPER_USER
+  persistentUserId: string | null
   constructor(input: Input) {
     const {
       tms,
@@ -55,7 +57,8 @@ export default class User {
       isWatched,
       preferredName,
       pseudoId,
-      sendSummaryEmail
+      sendSummaryEmail,
+      persistentUserId
     } = input
     const now = new Date()
     this.id = id ?? `local|${generateUID()}`
@@ -72,5 +75,6 @@ export default class User {
     this.preferredName = preferredName.trim().slice(0, USER_PREFERRED_NAME_LIMIT)
     this.pseudoId = pseudoId ?? undefined
     this.sendSummaryEmail = sendSummaryEmail ?? true
+    this.persistentUserId = persistentUserId ?? null
   }
 }